Mobe the fake docker tags and fake service to their own files
