+++ tests/Support/FakeDockerTags.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Tests\Support;
+
+class FakeDockerTags
+{
+    public function resolveTag($tag)
+    {
+        return $tag;
+    }
+}
